requirements.txt
streamlit==1.28.0
pinecone-client==2.2.4
groq==0.3.1
sentence-transformers==2.2.2
numpy==1.24.3
python-dotenv==1.0.0
